+++ utils/starter/go.mod
@@ -1,0 +1,3 @@
+module github.com/cosmos/sdk-tutorials/utils/starter
+
+go 1.12
